Financial_ai_app/requirements.txt
fastapi
uvicorn
sqlalchemy
PyJWT
sentence-transformers
faiss-cpu
numpy
PyPDF2